refac: Clean up files and folder structure of batches infrastructure project (#1190)
