+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/Actors/ActorRelation.cs
@@ -12,12 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient;
+namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.Actors;
 
-public class DatabricksSqlException : Exception
-{
-    public DatabricksSqlException(string message)
-        : base(message)
-    {
-    }
-}
+/// <summary>
+/// Class used for deserializing a single entry, in the actor set generated by Databricks.
+/// All values of the Actor must be in lower casing to conform with the spark json output.
+/// </summary>
+#pragma warning disable SA1300
+public sealed record ActorRelation(string energy_supplier_gln, string balance_responsible_party_gln);
