+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/CalculationResults/ProcessResultPoint.cs
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient.DeltaTableConstants;
+namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResults;
 
-public static class DeltaTableProcessType
-{
-    public const string Aggregation = "Aggregation";
-    public const string BalanceFixing = "BalanceFixing";
-}
+/// <summary>
+/// Class used for deserializing a single entry, in the result set generated by Databricks.
+/// All values of the PointDto must be in lower casing to conform with the spark json output.
+/// </summary>
+#pragma warning disable SA1300
+public sealed record ProcessResultPoint(string quantity, string quality, string quarter_time);
