+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/SqlStatements/DatabricksSqlResponseState.cs
@@ -12,11 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient;
+namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements;
 
 /// <summary>
-/// Class used for deserializing a single entry, in the result set generated by Databricks.
-/// All values of the PointDto must be in lower casing to conform with the spark json output.
+/// Representation of the state of a Databricks SQL response.
 /// </summary>
-#pragma warning disable SA1300
-public sealed record ProcessResultPoint(string quantity, string quality, string quarter_time);
+public enum DatabricksSqlResponseState
+{
+    Failed = 0,
+    Cancelled = 1,
+    Pending = 2,
+    Succeeded = 3,
+}
