+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/SqlStatements/IDatabricksSqlResponseParser.cs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.BatchActor;
+namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements;
 
-/// <summary>
-/// Class used for deserializing a single entry, in the actor set generated by Databricks.
-/// All values of the Actor must be in lower casing to conform with the spark json output.
-/// </summary>
-#pragma warning disable SA1300
-public sealed record ActorRelation(string energy_supplier_gln, string balance_responsible_party_gln);
+public interface IDatabricksSqlResponseParser
+{
+    DatabricksSqlResponse Parse(string jsonResponse);
+}
