+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/Application/ProcessStep/Model/ProcessResultPointTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient;
+using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResults;
 using Test.Core;
 using Xunit;
 using Xunit.Categories;
