+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/DeltaTableContracts/MeteringPointTypeMapperTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient.Mappers;
+using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.Mappers;
 using FluentAssertions;
 using Test.Core;
 using Xunit;
