+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/Infrastructure/Actor/ActorRelationTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.BatchActor;
+using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.Actors;
 using Test.Core;
 using Xunit;
 using Xunit.Categories;
