+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.UnitTests/Infrastructure/CalculationResultClient/Mappers/SettlementMethodMapperTest.cs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient.Mappers;
+using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.DeltaTableConstants;
+using Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.Mappers;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
 using FluentAssertions;
 using Xunit;
 using Xunit.Categories;
-using DeltaTableConstants = Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.CalculationResultClient.DeltaTableConstants;
 
 namespace Energinet.DataHub.Wholesale.CalculationResults.UnitTests.Infrastructure.CalculationResultClient.Mappers;
 
@@ -25,9 +25,9 @@ namespace Energinet.DataHub.Wholesale.CalculationResults.UnitTests.Infrastructur
 public class SettlementMethodMapperTests
 {
     [Theory]
-    [InlineData(DeltaTableConstants.DeltaTableTimeSeriesType.FlexConsumption, SettlementMethod.Flex)]
-    [InlineData(DeltaTableConstants.DeltaTableTimeSeriesType.Production, null)]
-    [InlineData(DeltaTableConstants.DeltaTableTimeSeriesType.NonProfiledConsumption, SettlementMethod.NonProfiled)]
+    [InlineData(DeltaTableTimeSeriesType.FlexConsumption, SettlementMethod.Flex)]
+    [InlineData(DeltaTableTimeSeriesType.Production, null)]
+    [InlineData(DeltaTableTimeSeriesType.NonProfiledConsumption, SettlementMethod.NonProfiled)]
     public void ToDeltaTableValue_ReturnsExpectedSettlementMethod(string timeSeriesType, SettlementMethod? expected)
     {
         // Act
